+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- Added `subTitle` to the `DataGridCell` component to display additional subtitle text in the row data.
 
 ### Fixed